requirements.txt
asgiref
certifi==2021.5.30
chardet==4.0.0
Django
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
djongo==1.3.6
dnspython==2.1.0
idna==2.10
Markdown==3.3.4
psycopg2
pymongo==3.11.4
pytz
requests==2.25.1
sqlparse==0.2.4
typing-extensions
urllib3==1.26.6
wincertstore==0.2
